backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1
psycopg2-binary==2.9.10
pgvector==0.3.6
sentence-transformers==3.3.1
tqdm==4.67.1
numpy==2.2.1
python-dotenv==1.0.1
openai==1.58.1
torch==2.5.1
rich>=13.0.0